[SPARK-14866][SQL] Break SQLQuerySuite out into smaller test suites
